pyproject.toml
[project]
name = "pangolin-headers"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.12.14",
    "fastapi>=0.116.1",
    "pyyaml>=6.0.2",
    "uvicorn>=0.35.0",
]
